Rename test files to be compliant with nose
