+++ thermod/tests/test_thermometer.py
