+++ thermod/tests/test_timetable.py
